# Conflicts:
#	src/concurrency/blank_controller.cpp
#	src/index/vectraflow.cpp

+++ src/storage/storage_manager.cpp
@@ -39,14 +39,14 @@ auto candy::StorageManager::getVectorsByUids(const std::vector<uint64_t>& vector
 auto candy::StorageManager::topk(const std::unique_ptr<VectorRecord>& record, int k) const -> std::vector<uint64_t> {
   const auto rec = record.get();
   std::priority_queue<std::pair<double, int32_t>> pq;
-  for (int i = 0; i < records_.size(); ++i) {
+  for (size_t i = 0; i < records_.size(); ++i) {
     const auto local_rec = records_[i].get();
     auto dist = engine_->EuclideanDistance(rec->data_, local_rec->data_);
-    if (pq.size() < k) {
-      pq.emplace(dist, i);
+    if (pq.size() < static_cast<size_t>(k)) {
+      pq.emplace(dist, static_cast<int32_t>(i));
     } else if (dist < pq.top().first) {
       pq.pop();
-      pq.emplace(dist, i);
+      pq.emplace(dist, static_cast<int32_t>(i));
     }
   }
   std::vector<uint64_t> result;
