+++ src/stream/data_stream_source/file_stream_source.cpp
@@ -31,7 +31,7 @@ void candy::FileStreamSource::Init() {
         auto elapsed = std::chrono::duration_cast<std::chrono::milliseconds>(
           std::chrono::steady_clock::now() - last_data_time
         ).count();
-        if (elapsed > timeout_ms_) {
+        if (static_cast<uint64_t>(elapsed) > timeout_ms_) {
           std::cout << "FileStreamSource timeout after " << elapsed << " ms" << std::endl;
           break;
         }
